% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_tmle}
\alias{extract_tmle}
\title{Extract TSM and ATE from a tmle object}
\usage{
extract_tmle(fit, obs_a, obs_y)
}
\arguments{
\item{fit}{a TMLE model from the \code{tmle} package}

\item{obs_a}{A vector of the observed binary exposures}

\item{obs_y}{A vector of the observed outcomes}
}
\value{
A list of class \code{tmleExtract} with elements \code{tsm} (the extract estimates) and \code{IF} a dataframe with the elements used to extract such estimates.
}
\description{
Extract TSM and ATE from a tmle object
}
