% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{print.tmleExtract}
\alias{print.tmleExtract}
\title{Print an object of class \code{tmleExtract}}
\usage{
\method{print}{tmleExtract}(x)
}
\arguments{
\item{x}{an object of class \code{tmleExtract}}
}
\description{
Print an object of class \code{tmleExtract}
}
