% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{tmle_extract}
\alias{tmle_extract}
\title{Extract TSM and ATE from a tmle object}
\usage{
tmle_extract(fit, obs_a, obs_y)
}
\arguments{
\item{fit}{a TMLE model from the \code{tmle} package}

\item{obs_a}{A vector of the observed binary exposures}

\item{obs_y}{A vector of the observed outcomes}
}
\value{
A list of class \code{tmleExtract} with elements \code{tsm} (the extract estimates) and \code{IF} a dataframe with the elements used to extract such estimates.
}
\description{
Extract TSM and ATE from a tmle object
}
\examples{
library(tmleExtract)
library(tmle)

set.seed(1)
n <- 250
W <- matrix(rnorm(n*3), ncol=3)
A <- rbinom(n,1, 1/(1+exp(-(.2*W[,1] - .1*W[,2] + .4*W[,3]))))
Y <- A + 2*W[,1] + W[,3] + W[,2]^2 + rnorm(n)
tmle_fit <- tmle(Y,A,W, Q.SL.library = "SL.glm", g.SL.library = "SL.glm")

tmle_extract(tmle_fit, A, Y)
}
